VectorToRasterTests
Added tests to show that vectorToRaster create cells outside the
polygon. I also noticed that if you are creating two rasters. One raster
with one polygon and another with two polygons (the same polygon and
another polygon that is beside that polygon). The cells over the first
polygon will not be the same in the two rasters.